Small fixes and signal processing functions (#309)
* display volume; speedy option

* return centre of mass

* flake

* get out indices

* coverage grid

* add mapping argument

* small bugfix in between sorted for inconsistent bounds

* coverage improvements

* raw data streamer: if cache corresponds to query, do not reload

* add fk and agc functions in the voltage module

* destripe: channel rejection prototype

* bugfix on destripe

* small bugfix in between sorted for inconsistent bounds

* coverage improvements

* fix orphan test
