+++ ibllib/io/spikeglx.py
@@ -39,6 +39,13 @@ class Reader:
         if self.is_mtscomp:
             self._raw = mtscomp.Reader()
             self._raw.open(self.file_bin, self.file_bin.with_suffix('.ch'))
+            if self._raw.shape != (self.ns, self.nc):
+                ftsec = self._raw.shape[0] / self.fs
+                _logger.warning(f"{sglx_file} : meta data and compressed chunks dont checkout\n"
+                                f"File duration: expected {self.meta['fileTimeSecs']},"
+                                f" actual {ftsec}\n"
+                                f"Will attempt to fudge the meta-data information.")
+                self.meta['fileTimeSecs'] = ftsec
         else:
             if self.nc * self.ns * 2 != self.nbytes:
                 ftsec = self.file_bin.stat().st_size / 2 / self.nc / self.fs
